pyproject.toml
[project]
name = "urotheliome-dash"
version = "0.1.0"
description = "Dashboard for the Urotheliome research project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dash==3.2.0",
    "dash-bio==1.0.2",
    "dash-bootstrap-components==2.0.4",
    "dash-mantine-components==2.1.0",
    "flask-sso",
    "gunicorn==23.0.0",
    "numpy==2.2.4",
    "pandas==2.3.2",
    "plotly==6.3.0",
    "python-dotenv==1.1.1",
    "scipy==1.16.1",
]
